WIP: support deep nesting of with_tag blocks
The code appears to have been originally designed to support deep
nesting, but the approach probably can't work with the design of RSpec
and the code was slightly mangled to work while sacrificing deep
nesting. This commit de-mangles it but exposes why the code can't work
with RSpec, with a couple of extra tests to help prove the point.