Fix deep nesting using with_tag blocks
The code appears to have been originally designed to narrow the scope of
the selectors when using nested with_tags. However this doesn't work the
way it was (probably) implemented and the code appears to have been
slightly mangled to make nesting work without scope narrowing. This
commit attempts to de-mangle the code a bit, but at the same time
introduces tests that show that this doesn't work (because while the
scope can be narrowed, it isn't un-narrowed afterwards).